+++ Servers/QuazalServer/RDVServices/GameServices/PS3DriverServices/MissionService.cs
@@ -1,11 +1,10 @@
 using QuazalServer.RDVServices.DDL.Models;
 using QuazalServer.QNetZ.Attributes;
 using QuazalServer.QNetZ.Interfaces;
-using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3DriverServices
 {
-    [RMCService(RMCProtocolId.MissionService)]
+    [RMCService((ushort)RMCProtocolId.MissionService)]
     public class MissionService : RMCServiceBase
     {
         [RMCMethod(3)]
