+++ Servers/QuazalServer/RDVServices/GameServices/PS3DriverServices/NotificationEventManager.cs
@@ -1,11 +1,10 @@
 using QuazalServer.RDVServices.DDL.Models;
 using QuazalServer.QNetZ.Attributes;
 using QuazalServer.QNetZ.Interfaces;
-using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3DriverServices
 {
-    [RMCService(RMCProtocolId.NotificationEventManager)]
+    [RMCService((ushort)RMCProtocolId.NotificationEventManager)]
     public class NotificationEventManager : RMCServiceBase
     {
         [RMCMethod(1)]
